dont remake if exists

+++ ontology/ontology-maker.pl
@@ -15,7 +15,7 @@ foreach my $ont (@onts) {
     my $stage = "stage/$fn";
     my $tgt = "target/$fn";
     runcmd("wget --no-check-certificate $ont -O $stage.tmp && mv $stage.tmp $stage") unless -f $stage;
-    runcmd("owltools $stage --merge-imports-closure --extract-mingraph -o -f ttl --prefix OBO http://purl.obolibrary.org/obo/ $tgt.tmp && mv $tgt.tmp $tgt");
+    runcmd("owltools $stage --merge-imports-closure --extract-mingraph -o -f ttl --prefix OBO http://purl.obolibrary.org/obo/ $tgt.tmp && mv $tgt.tmp $tgt") unless -f $tgt;
 }
 
 runcmd("owltools target/*.owl --merge-support-ontologies  -o -f ttl --prefix OBO http://purl.obolibrary.org/obo/ merged.owl");
